+++ internal/command/check/command.go
@@ -30,6 +30,7 @@ func Cmd(cont *container.Container) *cobra.Command {
 		Short:   i18n.Get("CheckShort"),
 		Long:    i18n.Get("CheckLongDesc"),
 		Example: i18n.Get("CheckExample"),
+		Args:    cobra.NoArgs,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			// 检查 container 是否初始化
 			if cont == nil {